# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: User Service
  description: Service for manipulate user data
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /users/activate:
    post:
      tags:
      - users
      summary: activate users
      description: Activate users by emails
      operationId: users#activate
      parameters:
      - name: array
        in: body
        required: true
        schema:
          type: array
          items:
            type: string
            example: Error nostrum dolor illum veniam.
      responses:
        "200":
          description: OK response.
      schemes:
      - http
...